Implement basic rendering of the cpu display memory to the window.
Add a basic argbitrary color palette. Add a hardcoded delay to attempt to get
programs to run at a decent rate. Resize the window to be 32x32 to match the
size of the new cpu video memory.

At this point, snake is starting to work! I
don't have any input or any random number generation, but I can see the snake
move across the screen and the program ends when it hits an edge.